Added clipper.export.h file to make DLL/so files Added a sample Delphi app that tests the DLL Redisgned C++ Polytree structure to use inheritance not generics Polytree64ToPolytreeD function not longer required and removed (#292) Fixed bug where Polytree wasn't descaling results (#291)

+++ CPP/Utils/ClipFileSave.cpp
@@ -32,7 +32,7 @@ private:
   {
     ClearHaystack();
     stream.seekg(0, ios::end);
-    haystack_len = stream.tellg();
+    haystack_len = static_cast<size_t>(stream.tellg());
     stream.seekg(0, ios::beg);
     haystack_ = new char[haystack_len];
     stream.read(haystack_, haystack_len);
